# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/promotion/org/{organisationId}": {
      "get": {
        "operationId": "PromotionController_findAll",
        "parameters": [
          {
            "name": "organisationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      }
    },
    "/promotion/{id}": {
      "get": {
        "operationId": "PromotionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      },
      "patch": {
        "operationId": "PromotionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromotionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      },
      "delete": {
        "operationId": "PromotionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      }
    },
    "/promotion": {
      "post": {
        "operationId": "PromotionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromotionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      }
    }
  },
  "info": {
    "title": "Bedones Agents API",
    "description": "API backend pour le CRM social Bedones",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "CreatePromotionDto": {
        "type": "object",
        "properties": {
          "organisationId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "discountType": {
            "type": "string",
            "enum": [
              "PERCENTAGE",
              "FIXED_AMOUNT"
            ]
          },
          "discountValue": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "productIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "organisationId",
          "name"
        ]
      },
      "UpdatePromotionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "discountType": {
            "type": "string",
            "enum": [
              "PERCENTAGE",
              "FIXED_AMOUNT"
            ]
          },
          "discountValue": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "PAUSED",
              "EXPIRED"
            ]
          },
          "productIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}